bot/requirements.txt
python-telegram-bot[webhooks]==21.6
anthropic>=0.39.0
httpx>=0.27.0
python-dotenv>=1.0.0
